sketch and text edditing
made a scetch of what i want it to look like with an arrow and piece and look desings from a game i played on my chess board. i picked this as if black plays frist white wins and if white goes first its a staylmate based of me playing both sides at the same time.